--- a/requirements.txt
+++ b/requirements.txt
@@ -9,5 +9,5 @@ websocket_client
 gevent-websocket
 bencode.py
 coincurve
-python-bitcoinlib
 maxminddb
+sslcrypto