--- a/src/lib/sslcrypto/__init__.py
+++ b/src/lib/sslcrypto/__init__.py
@@ -1,0 +1,6 @@
+__all__ = ["aes", "ecc", "rsa"]
+
+try:
+    from .openssl import aes, ecc, rsa
+except OSError:
+    from .fallback import aes, ecc, rsa