--- a/src/lib/sslcrypto/fallback/__init__.py
+++ b/src/lib/sslcrypto/fallback/__init__.py
@@ -1,0 +1,3 @@
+from .aes import aes
+from .ecc import ecc
+from .rsa import rsa