--- a/src/lib/sslcrypto/fallback/rsa.py
+++ b/src/lib/sslcrypto/fallback/rsa.py
@@ -1,0 +1,6 @@
+class RSA:
+    def get_backend(self):
+        return "fallback"
+
+
+rsa = RSA()