@SwissConfederation 2019-12-02
Don't include this **until the license is not fixed!** 
Be warned! I going to release publicly ZeroNet under **GPLv3+** immediately after if this get merged!