@SwissConfederation 2019-12-02
Because I have a copy of ZeroNet where these libraries are not replaced and **I will not replace them! I will be forced to release ZeroNet under GPLv3+ immediately!** 
@HelloZeroNet do not merge this until the licensing issue not solved! Also note that most code is not originates from @imachug he copied most of the code from other people including from pastebin! 

The voting was a big show to get some time to write up this library.