@filips123 2019-12-02
> Also note that most code is not originates from @imachug he copied most of the code from other people including from pastebin!

Prove this.