@SwissConfederation 2019-12-02
I already proven that he is copied from others and so these library he claims his own is nothing more but a compilation and is can't be licenses under MIT either!