@filips123 2019-12-02
1. You can't include MIT license in MIT license? That's something new...

2. So you think it is not possible to write cryptography-related library because somebody already wrote it and now whole cryptography operations are copyrighted?