@SwissConfederation 2019-12-02
> now whole cryptography operations are copyrighted?

**Not only copyrighted but also patented!**