@filips123 2019-12-02
And you own that patents, right?

Why do we already have more than one cryptography library?