@SwissConfederation 2019-12-02
> And you own that patents, right?

No but I can use patented works as long I don't claim it to be mine! 

Why do we already have more than one cryptography library?

Why now before the license change need to change it? This still not solve the licensing issue! And even if all libraries replaced **I can publish my version of ZeroNet under GPLv3+ immediately!** 

**You are all risking here all your copyrights in the previously released ZeroNet versions!**