@SwissConfederation 2019-12-02
@HelloZeroNet **Do not merge this until the licensing issue is not resolved or you seriously damage the ZeroNet project!** Merging this will give immediately permission for all forks to use either GPLv3+ or GPLv3-only on ZeroNet!