@filips123 2019-12-02
> No but I can use patented works as long I don't claim it to be mine!

Where does @imachug claims that all code is his? Where some code was taken from other people, this is clearly stated in that file.

And still, I don't know what is the difference with other cryptography libraries...

> This still not solve the licensing issue!

Yes it does. If incompatible libraries are replaced, incompatibilities are fixed.

> I can publish my version of ZeroNet under GPLv3+ immediately!

No you can't. You need agreement from contributors for that.

> Merging this will give immediately permission for all forks to use either GPLv3+ or GPLv3-only on ZeroNet!

WTF are you talking?