@SwissConfederation 2019-12-02
Replacing the library in order to avoid changing the license **will not replace the library in my copy and other 1,900 forks!** 

Under GPL I have already granted all copyright permission to modify and distribute the software under GPL but legally I can't distribute ZeroNet under GPLv2! I'm guaranteed the right to distribute the software with or without my modifications and this right is not revocable! 

This means if you replace the library, I'm getting a green light to release ZeroNet under a **compatible** GPL license of my choosing, which is GPLv3+!