@SwissConfederation 2019-12-02
I don't need a shit from contributors! I already granted all rights under GPL! 
You need to tell contributors that they are fooled! I urge you @HelloZeroNet to don't even think about to merge this until the license in not fixed or you get into serious trouble!