@SwissConfederation 2019-12-02
@imachug 
> Third, it's licensed under MIT (I can add any other license you choose) so we wouldn't have problems if we wanted to re-license a part of ZeroNet code under some lax license.

You can't re-license **part** of ZeroNet. Under GPL section 2 the **entire ZeroNet** is already released under GPL! 
You also can't claim ownership over your contributions to a work which is licensed under GPL! 
There is no such thing as relicensing your lines what you added to the source code!