@SwissConfederation 2019-12-02
> ..pybitcointools because it fixes a vulnerability (I don't want to name it here to avoid possible private key leaks).

You are a joke! Now you telling me that `pybitcointools` is vulnerable to private key leaks? I want to see what they will say in BitcoinTalk about this! :rofl: **Theymos** will be happy to hear it!