@imachug 2019-12-03
> Replacing the library in order to avoid changing the license will not replace the library in my copy and other 1,900 forks!

Other forks would be illegal but at least the official one would be okay.