@imachug 2019-12-03
> You can't re-license part of ZeroNet

I can't, but nofish can relicense some of his code. Not as if it's related to this PR... 

> Theymos will be happy to hear it!

Who's that guy? AFAIK all/most pybitcointools forks are abandoned.