@filips123 2019-12-03
> Replacing the library in order to avoid changing the license will not replace the library in my copy and other 1,900 forks!

Changing the license to GPLv3+ would magically replace it in forks?

> Under GPL I have already granted all copyright permission to modify and distribute the software under GPL but legally I can't distribute ZeroNet under GPLv2!

You have right to distribute it under GPLv2, not any GPL license. That's why it includes GPLv2 license file, not just GPL license file.

> This means if you replace the library, I'm getting a green light to release ZeroNet under a compatible GPL license of my choosing, which is GPLv3+!

Replacing the library means that (almost) all incompatibilities are fixed and if you want to comply with licenses, you need to download new libraries.

> I don't need a shit from contributors! I already granted all rights under GPL!

As FSF said, contributions only granted distribution under GPLv2.

> You need to tell contributors that you are fooled them!

Because we allowed them to do what they should be able to do?

> Under GPL section 2 the entire ZeroNet is already released under GPL!

No it is not. It is licensed under GNU General Public License **version 2**.

> You also can't claim ownership over your contributions to a work which is licensed under GPL!

You "care" so much about contributions and then you would ignore their opinion and agreement when changing the license...