@SwissConfederation 2019-12-03
@filips123 I never seen more retarded person in my life than you! I hope @imachug at least have some brain and can translate this to you. 

- **NOBODY CAN DISTRIBUTE ZERONET USING GPLv2, THAT WOULD BE ILLEGAL AND CRIMINAL!** 
- **CONTRIBUTORS ALREADY GRANTED ALL RIGHTS TO KEEP DISTRIBUTING ZERONET UNDER GPL!** 
- **THIS RIGHTS TO SHARE AND MODIFY ARE NOT REVOCABLE!** 

Because, I do have **irrevocable** rights to share the code and I can't do it legally under GPLv2, I'm obligated by law to share ZeroNet under **GPLv3+**! 

Contributors agreed to release the **entire** ZeroNet under GPL! Under section two of the GPL the **entire** ZeroNet is under the same license, contributors can't change the license on they portion of the code and certainly they can't claim proprietary rights!