@SwissConfederation 2019-12-03
> Who's that guy? AFAIK all/most pybitcointools forks are abandoned.

**Copyright is protected for 70 years!**