@HelloZeroNet 2019-12-03
Thanks, looking great. I will do some benchmark on it. To add new dependency we need to improve the updater script that check if the user has installed the required libs, notify is something is missing and also need to find a solution for macOS where the runtime is not updated right now.
As a temporary solution we can include this package to src/lib.