@HelloZeroNet 2019-12-09
Thanks, it's looking great and fast, my results on Windows:

```

OpenSSL 1.0.2s  28 May 2019 (currently bundled)

Old (2.94s):
* Running testCryptEciesEncrypt x 100 .......... Done in 0.716s = Fine (1.68x)
* Running testCryptEciesDecrypt x 500 .......... Done in 1.044s = Fine (1.25x)
* Running testCryptEciesDecryptMulti x 5 x 100 (2 threads) ..... Done in 0.707s = OK (0.96x)
* Running testCryptAesEncrypt x 10000 .......... Done in 0.276s = OK (0.98x)
* Running testCryptAesDecrypt x 10000 .......... Done in 0.201s = Fine (1.25x)

New (1.96s):
* Running testCryptEciesEncrypt x 100 .......... Done in 0.387s = WOW (3.10x)
* Running testCryptEciesDecrypt x 500 .......... Done in 0.695s = Fast (1.87x)
* Running testCryptEciesDecryptMulti x 5 x 100 (2 threads) ..... Done in 0.500s = Fine (1.36x)
* Running testCryptAesEncrypt x 10000 .......... Done in 0.223s = Fine (1.21x)
* Running testCryptAesDecrypt x 10000 .......... Done in 0.162s = Fine (1.55x)


OpenSSL 1.1.0j  20 Nov 2018 (bundled with Python 3.7)

Old (6.63s):
* Running testCryptEciesEncrypt x 100 .......... Done in 1.698s = Goodish (0.71x)
* Running testCryptEciesDecrypt x 500 .......... Done in 2.502s = Ehh (0.52x)
* Running testCryptEciesDecryptMulti x 5 x 100 (2 threads) ..... Done in 1.921s = Ehh (0.35x)
* Running testCryptAesEncrypt x 10000 .......... Done in 0.300s = OK (0.90x)
* Running testCryptAesDecrypt x 10000 .......... Done in 0.214s = OK (1.17x)

New (3.73s):
* Running testCryptEciesEncrypt x 100 .......... Done in 0.888s = Fine (1.35x)
* Running testCryptEciesDecrypt x 500 .......... Done in 1.432s = OK (0.91x)
* Running testCryptEciesDecryptMulti x 5 x 100 (2 threads) ..... Done in 0.942s = Goodish (0.72x)
* Running testCryptAesEncrypt x 10000 .......... Done in 0.264s = OK (1.02x)
* Running testCryptAesDecrypt x 10000 .......... Done in 0.213s = OK (1.17x)


OpenSSL 1.1.1d  10 Sep 2019 (bundled with Python 3.8)

Old (2.55s):
* Running testCryptEciesEncrypt x 100 .......... Done in 0.610s = Fast (1.97x)
* Running testCryptEciesDecrypt x 500 .......... Done in 0.894s = Fine (1.45x)
* Running testCryptEciesDecryptMulti x 5 x 100 (2 threads) ..... Done in 0.558s = Fine (1.22x)
* Running testCryptAesEncrypt x 10000 .......... Done in 0.285s = OK (0.95x)
* Running testCryptAesDecrypt x 10000 .......... Done in 0.209s = OK (1.20x)

New (1.99s):
* Running testCryptEciesEncrypt x 100 .......... Done in 0.419s = WOW (2.86x)
* Running testCryptEciesDecrypt x 500 .......... Done in 0.683s = Fast (1.90x)
* Running testCryptEciesDecryptMulti x 5 x 100 (2 threads) ..... Done in 0.452s = Fine (1.50x)
* Running testCryptAesEncrypt x 10000 .......... Done in 0.241s = OK (1.12x)
* Running testCryptAesDecrypt x 10000 .......... Done in 0.204s = Fine (1.22x)  
```

So it's 25-100% faster, than the current lib we use.

As for improvement the result of sslcrypto.ecc.get_backend() could include the path of the loaded openssl dll/so file.